# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Brinecrypt API
  description: |
    Brinecrypt is a secure resource management and encryption service designed to provide a centralized and secure way to store and manage sensitive configuration data, secrets, and other resources.

    ### Core Concepts

    *   **Resources:** Named entities stored within a **Namespace**. Resources can be of type `cleartext` (stored as is) or `encrypted` (transparently encrypted/decrypted).
    *   **Namespaces:** Logical groupings for resources. The `_` namespace is reserved for internal system resources like users.
    *   **Versions:** Every time a resource is updated via `PUT`, a new version is created. Brinecrypt maintains a history of all versions.
    *   **Principals:** Entities that can be granted permissions. These include **Users** and **Service Accounts** (SAs).
    *   **Permissions:** Granular access control based on a `ResourcePattern` (e.g., `prod/*`), a `Verb` (`list`, `read`, `write`, `delete`), and an optional expiry time.

    ### Authentication & Authorization

    Brinecrypt supports multiple authentication methods:
    1.  **Session Tokens:** Obtained via `/auth/login`. Used for interactive sessions.
    2.  **Personal Access Tokens (PATs):** Long-lived tokens issued by users for programmatic access.
    3.  **Capability Tokens:** Restricted tokens issued with a specific set of permissions, useful for delegation or narrow-scoped tasks.
...

...

    ### Encryption

...
  version: 1.0.0
servers:
...
paths:
...
  /api/v1/{namespace}:
    get:
      tags:
        - Resources
      summary: List resources in a namespace
      operationId: listResourcesInNamespace
      security:
        - bearerAuth: []
      parameters:
        - name: namespace
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of resource summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResourceSummary'
        '401':
          description: Unauthorized
  /api/v1/{namespace}/{name}:
    get:
      tags:
        - Resources
      summary: Get a resource by name (latest version)
      operationId: getResource
      security:
        - bearerAuth: []
      parameters:
        - name: namespace
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Resource details including the latest value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        '401':
          description: Unauthorized
        '404':
          description: Resource not found
    put:
      tags:
        - Resources
      summary: Create or update a resource
      operationId: putResource
      security:
        - bearerAuth: []
      parameters:
        - name: namespace
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutResourceRequest'
      responses:
        '204':
          description: Resource successfully created or updated
        '400':
          description: Invalid request body
        '401':
          description: Unauthorized
    delete:
      tags:
        - Resources
      summary: Delete a resource
      operationId: deleteResource
      security:
        - bearerAuth: []
      parameters:
        - name: namespace
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Resource successfully deleted
        '401':
          description: Unauthorized
        '404':
          description: Resource not found
  /api/v1/{namespace}/{name}/versions:
    get:
      tags:
        - Resources
      summary: List all versions of a resource
      operationId: listResourceVersions
      security:
        - bearerAuth: []
      parameters:
        - name: namespace
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of resource version summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResourceValueSummary'
        '401':
          description: Unauthorized
        '404':
          description: Resource not found
  /api/v1/{namespace}/{name}/{version}:
    get:
      tags:
        - Resources
      summary: Get a specific version of a resource
      operationId: getResourceByVersion
      security:
        - bearerAuth: []
      parameters:
        - name: namespace
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: version
          in: path
          required: true
          schema:
            type: integer
            format: uint64
      responses:
        '200':
          description: Resource version details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceValue'
        '400':
          description: Invalid version number
        '401':
          description: Unauthorized
        '404':
          description: Resource or version not found
  /api/v1/uuid/{uuid}:
    get:
      tags:
        - Resources
      summary: Get a resource value by its unique UUID
      operationId: getResourceValueByUUID
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Resource value details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceValue'
        '401':
          description: Unauthorized
        '404':
          description: Resource value not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ResourceType:
      type: string
      enum: [undefined, cleartext, encrypted]
      description: Type of resource data storage.
    EncryptionAlgorithm:
      type: string
      enum: [undefined, aes-256-gcm]
      description: Algorithm used to encrypt resource data.
    Namespace:
      type: object
      properties:
        name:
          type: string
      required:
        - name
...
    Resource:
      type: object
      properties:
        name:
          type: string
        type:
          $ref: '#/components/schemas/ResourceType'
        created_at:
          type: string
          format: date-time
        created_by:
          type: string
        retired_at:
          type: string
          format: date-time
        namespace:
          $ref: '#/components/schemas/Namespace'
        value:
          $ref: '#/components/schemas/ResourceValue'
      required:
        - name
        - type
        - created_at
        - created_by
    ResourceValue:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        version:
          type: integer
        data:
          type: string
          description: The actual value content (decrypted if requested and authorized).
        encryption_algorithm:
          $ref: '#/components/schemas/EncryptionAlgorithm'
        encryption_key_id:
          type: integer
        encryption_key:
          $ref: '#/components/schemas/EncryptionKey'
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        retired_at:
          type: string
          format: date-time
      required:
        - uuid
        - version
        - data
        - encryption_algorithm
        - created_at
    EncryptionKey:
      type: object
      properties:
        encrypted_dek:
          type: string
        kek_version:
          type: integer
        created_at:
          type: string
          format: date-time
      required:
        - encrypted_dek
        - kek_version
        - created_at
    ResourceSummary:
      type: object
      properties:
        name:
          type: string
        type:
          $ref: '#/components/schemas/ResourceType'
        created_at:
          type: string
          format: date-time
        created_by:
          type: string
        retired_at:
          type: string
          format: date-time
      required:
        - name
        - type
        - created_at
        - created_by
    ResourceValueSummary:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        version:
          type: integer
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        retired_at:
          type: string
          format: date-time
        encryption_algorithm:
          $ref: '#/components/schemas/EncryptionAlgorithm'
      required:
        - uuid
        - version
        - created_at
        - encryption_algorithm
...
    PutResourceRequest:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/ResourceType'
        value:
          type: string
      required:
        - type
        - value